+++ packages/react-core/src/filters/Filter.js
@@ -1,7 +1,14 @@
 import { filterFunctions } from './FilterTypes';
+import { FiltersLogicalOperators } from '../operations/constants/FiltersLogicalOperators';
 
-function passesFilter(columns, filters, feature) {
-  return columns.every((column) => {
+const LOGICAL_OPERATOR_METHODS = {
+  [FiltersLogicalOperators.AND]: 'every',
+  [FiltersLogicalOperators.OR]: 'some'
+};
+
+function passesFilter(columns, filters, feature, filtersLogicalOperator) {
+  const method = LOGICAL_OPERATOR_METHODS[filtersLogicalOperator];
+  return columns[method]((column) => {
     const columnFilters = filters[column];
     const columnFilterTypes = Object.keys(columnFilters);
 
@@ -25,7 +32,11 @@ function passesFilter(columns, filters, feature) {
   });
 }
 
-export function buildFeatureFilter({ filters = {}, type = 'boolean' }) {
+export function buildFeatureFilter({
+  filters = {},
+  type = 'boolean',
+  filtersLogicalOperator = FiltersLogicalOperators.AND
+}) {
   const columns = Object.keys(filters);
 
   if (!columns.length) {
@@ -34,16 +45,16 @@ export function buildFeatureFilter({ filters = {}, type = 'boolean' }) {
 
   return (feature) => {
     const f = feature.properties || feature;
-    const featurePassesFilter = passesFilter(columns, filters, f);
+    const featurePassesFilter = passesFilter(columns, filters, f, filtersLogicalOperator);
 
     return type === 'number' ? Number(featurePassesFilter) : featurePassesFilter;
   };
 }
 
 // Apply certain filters to a collection of features
-export function applyFilters(features, filters) {
+export function applyFilters(features, filters, filtersLogicalOperator) {
   return Object.keys(filters).length
-    ? features.filter(buildFeatureFilter({ filters }))
+    ? features.filter(buildFeatureFilter({ filters, filtersLogicalOperator }))
     : features;
 }