+++ packages/react-widgets/src/models/CategoryModel.js
@@ -1,12 +1,21 @@
 import { Methods, executeTask } from '@carto/react-workers';
 
 export const getCategories = async (props) => {
-  const { column, operationColumn, joinOperation, operation, filters, dataSource } = props;
+  const {
+    column,
+    operationColumn,
+    operation,
+    joinOperation,
+    filters,
+    filtersLogicalOperator,
+    dataSource
+  } = props;
 
   return executeTask(dataSource, Methods.FEATURES_CATEGORY, {
     filters,
-    joinOperation,
+    filtersLogicalOperator,
     operation,
+    joinOperation,
     column,
     operationColumn: operationColumn || column
   });