+++ packages/react-workers/src/workers/features.worker.js
@@ -85,13 +85,19 @@ function getGeojsonFeatures({ viewport, geometry, uniqueIdProperty }) {
   postMessage({ result: true });
 }
 
-function getFormula({ filters, operation, column, joinOperation }) {
+function getFormula({
+  filters,
+  filtersLogicalOperator,
+  operation,
+  column,
+  joinOperation
+}) {
   let result = null;
 
   if (currentFeatures) {
     const targetOperation = aggregationFunctions[operation];
 
-    const filteredFeatures = getFilteredFeatures(filters);
+    const filteredFeatures = getFilteredFeatures(filters, filtersLogicalOperator);
 
     result = [{ value: targetOperation(filteredFeatures, column, joinOperation) }];
   }
@@ -99,11 +105,18 @@ function getFormula({ filters, operation, column, joinOperation }) {
   postMessage({ result });
 }
 
-function getHistogram({ filters, operation, column, ticks, joinOperation }) {
+function getHistogram({
+  filters,
+  filtersLogicalOperator,
+  operation,
+  column,
+  ticks,
+  joinOperation
+}) {
   let result = null;
 
   if (currentFeatures) {
-    const filteredFeatures = getFilteredFeatures(filters);
+    const filteredFeatures = getFilteredFeatures(filters, filtersLogicalOperator);
 
     result = histogram({
       data: filteredFeatures,
@@ -117,11 +130,18 @@ function getHistogram({ filters, operation, column, ticks, joinOperation }) {
   postMessage({ result });
 }
 
-function getCategories({ filters, operation, column, operationColumn, joinOperation }) {
+function getCategories({
+  filters,
+  filtersLogicalOperator,
+  operation,
+  column,
+  operationColumn,
+  joinOperation
+}) {
   let result = null;
 
   if (currentFeatures) {
-    const filteredFeatures = getFilteredFeatures(filters);
+    const filteredFeatures = getFilteredFeatures(filters, filtersLogicalOperator);
 
     const groups = groupValuesByColumn({
       data: filteredFeatures,
@@ -139,6 +159,7 @@ function getCategories({ filters, operation, column, operationColumn, joinOperat
 
 function getScatterPlot({
   filters,
+  filtersLogicalOperator,
   xAxisColumn,
   yAxisColumn,
   xAxisJoinOperation,
@@ -146,7 +167,7 @@ function getScatterPlot({
 }) {
   let result = [];
   if (currentFeatures) {
-    const filteredFeatures = getFilteredFeatures(filters);
+    const filteredFeatures = getFilteredFeatures(filters, filtersLogicalOperator);
     result = scatterPlot({
       data: filteredFeatures,
       xAxisColumns: [xAxisColumn].flat(),
@@ -161,6 +182,7 @@ function getScatterPlot({
 
 function getTimeSeries({
   filters,
+  filtersLogicalOperator,
   column,
   stepSize,
   operation,
@@ -170,7 +192,7 @@ function getTimeSeries({
   let result = [];
 
   if (currentFeatures) {
-    const filteredFeatures = getFilteredFeatures(filters);
+    const filteredFeatures = getFilteredFeatures(filters, filtersLogicalOperator);
 
     const groups = groupValuesByDateColumn({
       data: filteredFeatures,
@@ -190,6 +212,7 @@ function getTimeSeries({
 // See sorting details in utils/sorting.js
 function getRawFeatures({
   filters,
+  filtersLogicalOperator,
   limit = 10,
   page = 0,
   sortBy,
@@ -200,7 +223,7 @@ function getRawFeatures({
   let totalCount = 0;
 
   if (currentFeatures) {
-    data = applySorting(getFilteredFeatures(filters), {
+    data = applySorting(getFilteredFeatures(filters, filtersLogicalOperator), {
       sortBy,
       sortByDirection
     });
@@ -222,6 +245,6 @@ function applyPagination(features, { limit, page }) {
   return features.slice(limit * Math.max(0, page), limit * Math.max(1, page + 1));
 }
 
-function getFilteredFeatures(filters = {}) {
-  return _applyFilters(currentFeatures, filters);
+function getFilteredFeatures(filters = {}, filtersLogicalOperator) {
+  return _applyFilters(currentFeatures, filters, filtersLogicalOperator);
 }